+++ apps/fluent-tester/src/TestComponents/RadioGroupV1/RadioGroupV1Test.tsx
@@ -30,6 +30,13 @@ const radioGroupV1Sections: TestSection[] = [
   },
   ...Platform.select({
     android: [null],
+    native: [
+      {
+        name: 'RadioGroup with Label Subtext',
+        testID: RADIOGROUPV1_TESTPAGE,
+        component: SubtextRadioGroup,
+      },
+    ],
     default: [
       {
         name: 'RadioGroup with Label Subtext',