+++ packages/components/RadioGroup/src/legacy/__tests__/RadioButton.test.tsx
@@ -1,0 +1,12 @@
+import * as React from 'react';
+
+import * as renderer from 'react-test-renderer';
+
+import { RadioButton } from '../RadioButton';
+
+describe('RadioButton component tests', () => {
+  it('RadioButton default', () => {
+    const tree = renderer.create(<RadioButton buttonKey="key1" content="Default Button" />).toJSON();
+    expect(tree).toMatchSnapshot();
+  });
+});