+++ packages/dependency-profiles/src/index.js
@@ -4,15 +4,15 @@ module.exports = {
   "0.68": {
     "@fluentui-react-native/e2e-testing": {
       "name": "@fluentui-react-native/e2e-testing",
-      "version": "1.23.2"
+      "version": "1.24.0"
     },
     "@fluentui-react-native/tester": {
       "name": "@fluentui-react-native/tester",
-      "version": "0.144.0"
+      "version": "0.146.0"
     },
     "@fluentui-react-native/tester-win32": {
       "name": "@fluentui-react-native/tester-win32",
-      "version": "0.32.82"
+      "version": "0.32.87"
     },
     "@fluentui-react-native/avatar": {
       "name": "@fluentui-react-native/avatar",
@@ -64,7 +64,7 @@ module.exports = {
     },
     "@fluentui-react-native/menu": {
       "name": "@fluentui-react-native/menu",
-      "version": "1.8.1"
+      "version": "1.9.0"
     },
     "@fluentui-react-native/menu-button": {
       "name": "@fluentui-react-native/menu-button",
@@ -88,7 +88,7 @@ module.exports = {
     },
     "@fluentui-react-native/radio-group": {
       "name": "@fluentui-react-native/radio-group",
-      "version": "0.16.46"
+      "version": "0.16.47"
     },
     "@fluentui-react-native/separator": {
       "name": "@fluentui-react-native/separator",
@@ -162,13 +162,17 @@ module.exports = {
       "name": "@fluentui-react-native/experimental-checkbox",
       "version": "0.13.53"
     },
+    "@fluentui-react-native/drawer": {
+      "name": "@fluentui-react-native/drawer",
+      "version": "0.1.0"
+    },
     "@fluentui-react-native/dropdown": {
       "name": "@fluentui-react-native/dropdown",
       "version": "0.7.49"
     },
     "@fluentui-react-native/experimental-expander": {
       "name": "@fluentui-react-native/experimental-expander",
-      "version": "0.5.26"
+      "version": "0.5.27"
     },
     "@fluentui-react-native/experimental-link": {
       "name": "@fluentui-react-native/experimental-link",
@@ -192,7 +196,7 @@ module.exports = {
     },
     "@fluentui-react-native/experimental-radio-group": {
       "name": "@fluentui-react-native/experimental-radio-group",
-      "version": "0.9.34"
+      "version": "0.9.36"
     },
     "@fluentui-react-native/experimental-shadow": {
       "name": "@fluentui-react-native/experimental-shadow",
@@ -264,7 +268,7 @@ module.exports = {
     },
     "@fluentui/react-native": {
       "name": "@fluentui/react-native",
-      "version": "0.36.25"
+      "version": "0.36.26"
     },
     "@fluentui-react-native/android-theme": {
       "name": "@fluentui-react-native/android-theme",