+++ packages/experimental/RadioGroup/src/index.ts
@@ -1,16 +1,31 @@
-export { RadioGroup } from './RadioGroup/RadioGroup';
+if (__DEV__) {
+  console.warn(
+    'The @fluentui-react-native/exprimental-radio-group package is deprecated. The contents of this package have been moved to @fluentui-react-native/radio-group. If you need to use the RadioGroup component from this package, please use RadioGroupV1 from @fluentui-react-native/radio-group.',
+  );
+}
+
 export {
-  radioGroupName,
+  RadioGroupV1 as RadioGroup,
+  radioGroupNameV1 as radioGroupName,
   RadioGroupInfo,
   RadioGroupProps,
   RadioGroupSlotProps,
   RadioGroupState,
   RadioGroupTokens,
   RadioGroupType,
-} from './RadioGroup/RadioGroup.types';
-export { RadioGroupContext, RadioGroupContextValue, RadioGroupProvider, useRadioGroupContext } from './RadioGroup/radioGroupContext';
-export { useRadioGroup } from './RadioGroup/useRadioGroup';
-export { useRadioGroupContextValue } from './RadioGroup/useRadioGroupContextValue';
-export { Radio, radioLookup } from './Radio/Radio';
-export { radioName, RadioProps, RadioSlotProps, RadioInfo, RadioTokens, RadioType } from './Radio/Radio.types';
-export { useRadio } from './Radio/useRadio';
+  RadioGroupContextV1 as RadioGroupContext,
+  RadioGroupContextValue,
+  RadioGroupProvider,
+  useRadioGroupContext,
+  useRadioGroup,
+  useRadioGroupContextValue,
+  Radio,
+  radioLookup,
+  radioName,
+  RadioProps,
+  RadioSlotProps,
+  RadioInfo,
+  RadioTokens,
+  RadioType,
+  useRadio,
+} from '@fluentui-react-native/radio-group';