+++ packages/experimental/Expander/CHANGELOG.md
@@ -1,8 +1,16 @@
 # Change Log - @fluentui-react-native/experimental-expander
 
-This log was last generated on Wed, 03 May 2023 07:08:52 GMT and should not be manually modified.
+This log was last generated on Wed, 17 May 2023 07:07:43 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.5.27
+
+Wed, 17 May 2023 07:07:43 GMT
+
+### Patches
+
+- bump expander's windows min (tatianakapos@microsoft.com)
 
 ## 0.5.26